Feature/nav (#118)
* Move navigation stacks into own file and move index state to app

* Move navigation components into own files

* Wrap FormAgent story with navigation

* Remove boilerplate app file in root

* Remove YellowBox warning configs from App.js

* Clean up

* Add working dummy test

* Clean up
