+++ source/components/atoms/Badge/Badge.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components/native';
+
+const Badge = styled.View`
+  border-color: #f5f5f5;
+  border-width: 2px;
+  position: absolute;
+  right: -8px;
+  top: -10px;
+  background-color: #d73640;
+  border-radius: 24px;
+  width: 20px;
+  height: 20px;
+  justify-content: center;
+  align-items: center;
+`;
+
+export default Badge;
