+++ source/components/atoms/Badge/BadgeText.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components/native';
+
+const BadgeText = styled.Text`
+  color: white;
+  font-size: 12px;
+  font-family: Roboto;
+  font-weight: 800;
+`;
+
+export default BadgeText;
