+++ source/components/atoms/Badge/index.js
@@ -1,0 +1,2 @@
+export { default } from './Badge';
+export { default as BadgeText } from './BadgeText';
