+++ source/components/molecules/MaterialTopTabBarWrapper/index.js
@@ -1,0 +1,1 @@
+export { default } from './MaterialTopTabBarWrapper';
