+++ source/components/molecules/TabBarIcon/TabBarIcon.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Icon } from 'react-native-elements';
+// TODO: Move Badge to own component file
+const TabBarIcon = (iconName, colorFocused) => ({ focused }) => (
+  <Icon name={iconName} color={focused ? colorFocused : 'gray'} />
+);
+
+export default TabBarIcon;
