+++ source/components/molecules/TabBarIcon/index.js
@@ -1,0 +1,1 @@
+export { default } from './TabBarIcon';
